Refactor NostrEventProvider and move constants to a new file
