+++ providers/NostrEventProvider.tsx
@@ -17,103 +17,17 @@ import {
   getFollowsListMap,
   EventCache,
   getEventArray,
-  getParentEventId,
 } from "@/utils";
 import { useNostrRelayList } from "@/hooks/nostrRelayList";
 import { pool } from "@/utils/relay-pool";
 import { useAuth } from "@/hooks";
-
-const FORTY_EIGHT_HOURS = 48 * 60 * 60 * 1000;
-const SOCIAL_NOTES: Filter = {
-  kinds: [
-    // comment
-    1,
-    // repost
-    6,
-    // generic repost
-    16,
-    // reaction
-    7,
-    // zap receipt
-    9735,
-  ],
-  limit: 100,
-};
-
-const FOLLOWS_SOCIAL_NOTES: Filter = {
-  kinds: [
-    // comment
-    1,
-    // repost
-    6,
-    // generic repost
-    16,
-  ],
-  limit: 52,
-};
-
-const RECENT_COMMENT_ACTIVITY: Filter = {
-  kinds: [
-    // repost
-    6,
-    // generic repost
-    16,
-    // reaction
-    7,
-    // zap receipt
-    9735,
-  ],
-  limit: 53,
-};
-
-const PUBKEY_METADATA: Filter = {
-  kinds: [0],
-};
-
-const FOLLOWS_FILTER: Filter = {
-  kinds: [
-    // Follows
-    3,
-    // mute list
-    10000,
-    // user emoji list
-    10030,
-  ],
-};
-
-type NostrEventContextType = {
-  querySync: (filter: Filter) => Promise<Event[]>;
-  getEventAsync: (id: string) => Promise<Event | null>;
-  cacheEventById: (event: Event) => void;
-  cacheEventsById: (events: Event[]) => void;
-  getPubkeyProfile: (pubkey: string) => Promise<NostrUserProfile | null>;
-  batchGetPubkeyProfiles: (
-    pubkeys: string[],
-  ) => Promise<Map<string, NostrUserProfile>>;
-  getEventRelatedEvents: (event: Event) => Promise<Event[]>;
-  comments: Event[];
-  reactions: Event[];
-  reposts: Event[];
-  genericReposts: Event[];
-  zapReceipts: Event[];
-  follows: string[];
-  followsActivity: string[];
-  // Record<followPubkey, followRelay>
-  followsMap: Record<string, string>;
-  loadInitialData: () => Promise<void>;
-  isLoadingInitial: boolean;
-};
-
-const defaultNostrEventContext: Partial<NostrEventContextType> = {
-  comments: [],
-  reactions: [],
-  reposts: [],
-  genericReposts: [],
-  zapReceipts: [],
-  follows: [],
-  followsActivity: [],
-  followsMap: {},
-};
+import {
+  FORTY_EIGHT_HOURS,
+  getQueryKeyForKind,
+  NostrEventContextType,
+  nostrQueryKeys,
+  SOCIAL_NOTES,
+} from "./constants";
 
 const NostrEventContext = createContext<NostrEventContextType | null>(null);
 
@@ -414,42 +328,6 @@ export function useNostrEvents() {
   return context;
 }
 
-export const nostrQueryKeys = {
-  event: (id: string) => ["nostr", "event", id],
-  profile: (pubkey: string) => ["nostr", "profile", "event", pubkey],
-  relayList: (pubkey: string) => ["nostr", "relayList", "event", pubkey],
-  // TODO - clean up old follows hooks (add, remove, get)
-  follows: (pubkey: string | null | undefined) => [
-    "nostr",
-    "follows",
-    "event",
-    pubkey,
-  ],
-  // kind 1 with user pubkey #p tag
-  comments: (pubkey: string) => ["nostr", "kind1", pubkey],
-  // kind 6 with user pubkey #p tag
-  reposts: (pubkey: string) => ["nostr", "kind6", pubkey],
-  // kind 7 with user pubkey #p tag
-  reactions: (pubkey: string) => ["nostr", "kind7", pubkey],
-  // kind 16 with user pubkey #p tag
-  genericReposts: (pubkey: string) => ["nostr", "kind16", pubkey],
-  // kind 9735 with user pubkey #p tag
-  zapReceipts: (pubkey: string) => ["nostr", "kind9735", pubkey],
-  // any kind with event #e tag that references the eventId (replies, reposts, reactions, etc)
-  eventRelatedEvents: (eventId: string) => [
-    "nostr",
-    "eventRelatedEvents",
-    eventId,
-  ],
-  contentComments: (contentId: string) => [
-    "nostr",
-    "content-comments",
-    contentId,
-  ],
-  // kind 1 with event #e tag
-  replies: (eventId: string) => ["nostr", "replies", eventId],
-};
-
 const decodeProfileMetadata = (event: Event) => {
   try {
     return JSON.parse(event.content) as NostrUserProfile;
@@ -493,18 +371,3 @@ const getEventSince = async (filter: Filter, readRelayList: string[]) => {
   await setLastFetchTime(queryKey);
   return event;
 };
-
-function getQueryKeyForKind(kind: number, pubkey: string) {
-  switch (kind) {
-    case 1:
-      return nostrQueryKeys.comments(pubkey);
-    case 6:
-      return nostrQueryKeys.reposts(pubkey);
-    case 7:
-      return nostrQueryKeys.reactions(pubkey);
-    case 16:
-      return nostrQueryKeys.genericReposts(pubkey);
-    case 9735:
-      return nostrQueryKeys.zapReceipts(pubkey);
-  }
-}
